refactor: simplify Python error handling to focus on assertions only
- Remove over-engineered context detection from pythonUtils
- Delete generic pythonError module from python/ directory
- Keep only assertion-specific error formatting in assertions/pythonError.ts
- Remove unnecessary file validation (Python will handle this)
- Use dynamic import to avoid circular dependencies

This is a much simpler approach that:
1. Only enhances error messages where we have clear context (assertions)
2. Doesn't try to guess context from method names
3. Lets each integration handle its own errors appropriately
4. Reduces complexity and maintenance burden
