--- test/python-errors/assertion-error.yaml
@@ -1,42 +1,0 @@
-# Test Python assertion errors
-
-prompts:
-  - "Test prompt"
-
-providers:
-  - openai:gpt-4o-mini
-
-tests:
-  - vars:
-      expected: "test"
-    assert:
-      # Test missing module error
-      - type: python
-        value: |
-          import numpy as np
-          def get_assert(output, context):
-              return True
-      
-      # Test function signature error
-      - type: python
-        value: |
-          def get_assert(output):  # Missing context parameter
-              return True
-      
-      # Test invalid return value
-      - type: python
-        value: |
-          def get_assert(output, context):
-              return "invalid"  # Should return bool, float, or dict
-      
-      # Test missing file:// prefix
-      - type: python
-        value: "missing_file.py"
-      
-      # Test missing file:// prefix with function
-      - type: python
-        value: "missing_file.py:custom_assert"
-      
-      # Test file not found
-      - type: python
-        value: "file://nonexistent_assert.py"
