--- test/python-errors/prompt-error.yaml
@@ -1,11 +1,0 @@
-# Test Python prompt errors
-
-prompts:
-  - file://test/python-errors/prompt-error.py
-
-providers:
-  - openai:gpt-4o-mini
-
-tests:
-  - vars:
-      test: "value"
