--- test/python-errors/transform-error.yaml
@@ -1,14 +1,0 @@
-# Test Python transform errors
-
-prompts:
-  - "Test prompt"
-
-providers:
-  - openai:gpt-4o-mini
-
-defaultTest:
-  transform: "file://test/python-errors/transform-error.py"
-
-tests:
-  - vars:
-      test: "value"
